+++ packages/commercetools/api-client/src/fragments/inventory-entries-query-result.ts
@@ -1,0 +1,14 @@
+import { InventoryEntryFragment } from './inventory';
+
+export const InventoryEntriesQueryResultFragment = `
+  ${InventoryEntryFragment}
+
+  fragment InventoryEntriesQueryResultFragment on InventoryEntryQueryResult {
+    offset
+    count
+    total
+    results {
+      ...InventoryEntryFragment
+    }
+  }
+`;