+++ packages/commercetools/composables/src/types/index.ts
@@ -1,4 +1,11 @@
-import { Filter, ProductVariant, Category, ResetPasswordResponse, CreatePasswordResetTokenResponse } from '@vue-storefront/commercetools-api';
+import {
+  Filter,
+  ProductVariant,
+  Category,
+  ResetPasswordResponse,
+  CreatePasswordResetTokenResponse,
+  Store
+} from '@vue-storefront/commercetools-api';
 import { FacetSearchResult } from '@vue-storefront/core';
 import { StoreQueryResult } from './GraphQL';
 
@@ -37,5 +44,10 @@ export interface ForgotPasswordResult {
 export type SearchData = FacetSearchResult<FacetResultsData>
 
 export interface StoresData extends StoreQueryResult {
-  _selected: string;
+  _selectedStore: string;
+}
+
+export interface StoreGetters<STORES, CRITERIA = any> {
+  getItems(stores: STORES, criteria?: CRITERIA): Store[];
+  getSelected(stores: STORES): Store | undefined;
 }