+++ packages/core/docs/.vuepress/config.js
@@ -31,6 +31,14 @@ module.exports = {
 		sidebar: {
 			'/migrate/': [
 				{
+					title: 'Migration guide 2.4.0',
+					children: [
+						['/migrate/2.4.0/overview', 'Overview'],
+						['/migrate/2.4.0/integrators', 'Integrators'],
+						['/migrate/2.4.0/commercetools', 'commercetools']
+					]
+				},
+				{
 					title: 'Migration guide 2.3.0',
 					children: [
 						['/migrate/2.3.0/overview', 'Overview'],
@@ -159,6 +167,7 @@ module.exports = {
 						['/shopify/', 'Introduction'],
 						['/shopify/getting-started', 'Getting Started'],
 						['/shopify/configuration', 'Configuration'],
+						['/shopify/authorization-strategy', 'Authorization'],
 						['/shopify/feature-list', 'Feature list'],
 						['/shopify/maintainers', 'Maintainers and support']
 					]
@@ -167,15 +176,30 @@ module.exports = {
 					title: 'Composables',
 					collapsable: false,
 					children: [
+						['/shopify/use-product', 'useProduct'],
+						['/shopify/use-user', 'useUser'],
+						['/shopify/use-user-order', 'useUserOrder'],
 						['/shopify/use-cart', 'useCart'],
 						['/shopify/use-category', 'useCategory'],
 						['/shopify/use-content', 'useContent'],
-						['/shopify/use-product', 'useProduct'],
 						['/shopify/use-search', 'useSearch'],
-						['/shopify/use-user', 'useUser'],
-						['/shopify/use-user-order', 'useUserOrders'],
 					]
-				}, {
+				},
+				{
+					title: 'API Client',
+					collapsable: false,
+					children: [
+						['/shopify/api-client-reference', 'Methods reference']
+					]
+				},
+				{
+					title: 'Theme',
+					collapsable: false,
+					children: [
+						['/shopify/auth-middleware', 'Auth Middleware']
+					],
+				},
+				{
 					title: 'Other',
 					collapsable: false,
 					children: [