+++ packages/core/docs/changelog/5924.js
@@ -1,8 +1,0 @@
-module.exports = {
-  description: 'cart item qty selector disabled on loading',
-  link: 'https://github.com/vuestorefront/vue-storefront/pull/5924',
-  isBreaking: false,
-  breakingChanges: [],
-  author: 'vn-vlad',
-  linkToGitHubAccount: 'https://github.com/vn-vlad'
-};