+++ packages/core/docs/changelog/6122.js
@@ -1,8 +1,0 @@
-module.exports = {
-  description: 'Updates form validation scheme for street, number and city in the checkout and profile editing pages',
-  link: 'https://github.com/vuestorefront/vue-storefront/pull/6122',
-  isBreaking: false,
-  breakingChanges: [],
-  author: 'Heitor Ramon Ribeiro',
-  linkToGitHubAccount: 'https://github.com/bloodf'
-};