+++ packages/core/docs/commercetools/changelog/6050.js
@@ -1,8 +1,0 @@
-module.exports = {
-  description: 'Refactor updateCart, fix retry logic in case of mismatch of cart',
-  link: 'https://github.com/vuestorefront/vue-storefront/pull/6054',
-  isBreaking: false,
-  breakingChanges: [],
-  author: 'git-antonyuk',
-  linkToGitHubAccount: 'https://github.com/git-antonyuk'
-};