+++ packages/rollup.base.config.js
@@ -1,6 +1,23 @@
+import nodeResolve from '@rollup/plugin-node-resolve';
 import typescript from 'rollup-plugin-typescript2';
+import { terser } from 'rollup-plugin-terser';
 
-export function generateBaseConfig(pkg) {
+const extensions = ['.ts', '.js'];
+
+export function generateBaseConfig(pkg, useTerser = false) {
+  const plugins = [
+    nodeResolve({
+      extensions
+    }),
+    typescript({
+      useTsconfigDeclarationDir: true,
+      // eslint-disable-next-line global-require
+      typescript: require('typescript')
+    })
+  ];
+
+  if (useTerser) plugins.push(terser());
+
   return {
     input: 'src/index.ts',
     output: [
@@ -19,11 +36,6 @@ export function generateBaseConfig(pkg) {
       ...Object.keys(pkg.dependencies || {}),
       ...Object.keys(pkg.peerDependencies || {})
     ],
-    plugins: [
-      typescript({
-        // eslint-disable-next-line global-require
-        typescript: require('typescript')
-      })
-    ]
+    plugins
   };
 }