+++ packages/core/docs/changelog/6198.js
@@ -1,8 +1,0 @@
-module.exports = {
-  description: 'Fix `@vue-storefront/cache` package',
-  link: 'https://github.com/vuestorefront/vue-storefront/pull/6198',
-  isBreaking: false,
-  breakingChanges: [],
-  author: 'Filip Sobol',
-  linkToGitHubAccount: 'https://github.com/filipsobol'
-};